Moves UntrustedInterfaces to a separate file
