+++ consensus/service/src/tx_manager/untrusted_interfaces.rs
@@ -1,0 +1,39 @@
+use mc_consensus_enclave::WellFormedTxContext;
+use mc_crypto_keys::CompressedRistrettoPublic;
+use mc_transaction_core::{
+    ring_signature::KeyImage,
+    tx::{TxHash, TxOutMembershipProof},
+    validation::TransactionValidationResult,
+};
+use std::sync::Arc;
+
+#[cfg(test)]
+use mockall::*;
+
+/// The untrusted (i.e. non-enclave) part of validating and combining transactions.
+#[cfg_attr(test, automock)]
+pub trait UntrustedInterfaces: Send + Sync {
+    /// Performs the untrusted part of the well-formed check.
+    /// Returns current block index and membership proofs to be used by
+    /// the in-enclave well-formed check on success.
+    fn well_formed_check(
+        &self,
+        highest_indices: &[u64],
+        key_images: &[KeyImage],
+        output_public_keys: &[CompressedRistrettoPublic],
+    ) -> TransactionValidationResult<(u64, Vec<TxOutMembershipProof>)>;
+
+    /// Checks if a transaction is valid (see definition in validators.rs).
+    fn is_valid(&self, context: Arc<WellFormedTxContext>) -> TransactionValidationResult<()>;
+
+    /// Combines a set of "candidate values" into a "composite value".
+    /// This assumes all values are well-formed and safe to append to the ledger individually.
+    ///
+    /// # Arguments
+    /// * `tx_contexts` - "Candidate" transactions. Each is assumed to be individually valid.
+    /// * `max_elements` - Maximal number of elements to output.
+    ///
+    /// Returns a bounded, deterministically-ordered list of transactions that are safe to append to the ledger.
+    fn combine(&self, tx_contexts: &[Arc<WellFormedTxContext>], max_elements: usize)
+        -> Vec<TxHash>;
+}
